Add files via upload
Traffic Light Signal files for all simulation scenarios. The content of each TLS file is already inserted in each net.xml-file respective. It was a safety measure designed to prevent SUMO from accidentally switch between TLS schemes. (If the TLS scheme is loaded as additional, the net.xml-file still has its built-in  default scheme.)
These additionals are authentic in such way that they still refer to both 'g' and 'G' for green signal. Before the final simulations, this was changed so only 'g' were used. The difference is that vehicles that get a 'g'-signal must yield to others, while vehicles that get a 'G'-signal has a higher priority.